Make anonsets double

+++ WalletWasabi.Gui/Controls/WalletExplorer/CoinViewModel.cs
@@ -183,7 +183,7 @@ namespace WalletWasabi.Gui.Controls.WalletExplorer
 
 		public uint OutputIndex => Model.Index;
 
-		public int AnonymitySet => Model.AnonymitySet;
+		public int AnonymitySet => (int)Model.AnonymitySet;
 
 		public string InCoinJoin => Model.CoinJoinInProgress ? "Yes" : "No";
